+++ src/logic/Token.cpp
@@ -1,7 +1,7 @@
 #include "Token.hpp"
 
 Token::Token(Army army, std::string name) : name(name){
-  this -> army = new Attribute("army", army);
+  this -> army = new Attribute("army", static_cast<int>(army));
 }
 
 Army Token::getArmy() {
